requirements.txt
requests
django==3.0.5
pytest==5.4.1
pytest-django
djangorestframework==3.11.0
djangorestframework-simplejwt

django-filter==2.2.0